fix: move static server and refactore
